backend/requirements-imsearch.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.2
faiss-cpu==1.9.0
fastapi==0.114.2
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.9.0
gradio==4.44.1
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.9
importlib_resources==6.4.5
iso8601==2.1.0
Jinja2==3.1.4
jiter==0.5.0
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numpy==2.1.1
openai==1.45.0
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
propcache==0.2.0
pydantic==2.9.1
pydantic_core==2.23.3
pydub==0.25.1
Pygments==2.18.0
pymongo==4.8.0
pyparsing==3.1.4
pypika-tortoise==0.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
ruff==0.6.9
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
sympy==1.13.1
tomlkit==0.12.0
tortoise-orm==0.21.7
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.30.6
websockets==12.0
yarl==1.17.0
